pyproject.toml
[project]
name = "heron"
version = "0.1.0"
description = "LLM-augmented algorithmic trading system"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.33.1",
    "feedparser>=6.0",
    "flask>=3.0",
    "httpx>=0.27",
    "pytz>=2024.1",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
heron = "heron.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["heron*"]
